requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
configobj==5.0.6
cssselect==1.1.0
decorator==5.1.1
esprima==4.0.1
fuzzywuzzy==0.18.0
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
javalang==0.13.0
lxml==4.6.3
pyquery==1.4.3
python-Levenshtein==0.12.2
requests==2.26.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
typing_extensions==4.0.1
urllib3==1.26.7
validators==0.18.2

mysql-connector-python~=8.0.29
Django~=4.0.6